backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic[email]==2.6.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pinecone
sentence-transformers==2.5.1
markdown==3.5.2
beautifulsoup4==4.13.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7 
pytest 